requirements.txt
absl-py==2.1.0
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.4
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.12
dlib==19.24.99
docker-pycreds==0.4.0
easydict==1.13
eval_type_backport==0.2.2
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.54.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.8.4
grpcio==1.67.0
huggingface-hub==0.26.1
humanfriendly==10.0
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
insightface==0.7.3
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
loguru==0.7.3
lpips==0.1.4
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
mxnet==1.9.1
networkx==3.2.1
ninja==1.11.1.1
numpy==1.23.1
onnx==1.17.0
onnxruntime==1.19.2
onnxruntime-gpu==1.19.2
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
prettytable==3.16.0
protobuf==5.28.3
psutil==6.1.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.1.4
PyQt5==5.15.10
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
sentry-sdk==2.19.0
setproctitle==1.3.4
simsimd==6.2.1
six==1.16.0
smmap==5.0.1
stringzilla==3.12.5
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.11
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1722737464726/work
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.2.3
validators==0.35.0
wandb==0.18.7
wcwidth==0.2.13
Werkzeug==3.0.4
yacs==0.1.8
zipp==3.20.2
